+++ src/python_template/contracts/__init__.py
@@ -1,0 +1,23 @@
+"""Contracts package public API."""
+
+from .protocols import (
+    AsyncCloseable,
+    AsyncFileReader,
+    AsyncFileWriter,
+    Closeable,
+    Configurable,
+    FileReader,
+    FileWriter,
+    Serializable,
+)
+
+__all__ = [
+    "Serializable",
+    "FileReader",
+    "AsyncFileReader",
+    "FileWriter",
+    "AsyncFileWriter",
+    "Closeable",
+    "AsyncCloseable",
+    "Configurable",
+]