+++ src/python_template/utils/date_utils.py
@@ -6,7 +6,7 @@
 import traceback
 from datetime import datetime, timedelta, timezone
 
-from .logger_util import get_logger
+from python_template.observability.log_config import get_logger
 
 logger = get_logger(__name__)